requirements.txt
ansi2html==1.8.0
blinker==1.7.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.10.0
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.0
fonttools==4.44.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
nest-asyncio==1.5.8
numpy==1.26.1
opencv-python-headless==4.8.0.74
packaging==23.2
Pillow==10.1.0
plotly==5.18.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
roboflow==1.1.9
scipy==1.11.3
six==1.16.0
supervision==0.16.0
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
